+++ CHANGELOG.rst
@@ -9,6 +9,7 @@ Unreleased
 Added
 -----
 * Visualization library, including plot_sma.
+* TO DO list.
 
 Changed
 -------